use distilbert as the backbone model for the text vectorization (#3)
